requirements.txt
fastapi
pydantic
uvicorn
python-decouple
python-multipart
Pillow
pyzbar
qrcode
pyngrok